Fixed v0.1.3 release data in change log
