+++ .jshintrc
@@ -1,0 +1,14 @@
+{
+  "bitwise": true,
+  "curly": true,
+  "eqeqeq": true,
+  "forin": true,
+  "freeze": true,
+  "indent": 2,
+  "latedef": true,
+  "quotmark": "double",
+  "strict": true,
+  "trailing": true,
+  "undef": true,
+  "unused": true
+}
