+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+v0.1.3
+-------------
+Release Date: 2014-05-02
+
+  * Bug fix for misnamed variable in _toArray() method (submitted by @danielmahal)
+  * Added gulp integration to lint, minify, and test code
+
 v0.1.2
 -------------
 Release Date: 2014-05-02
