+++ README.md
@@ -1,5 +1,4 @@
-ReactFire
-=========
+# ReactFire
 
 ReactFireMixin is an officially supported [ReactJS](http://facebook.github.io/react/) mixin
 for [Firebase](http://www.firebase.com/). Firebase provides your React app with a
@@ -7,8 +6,8 @@ persistent, realtime backend to effortlessly keep all of your clients in sync!
 
 Read our [blog post](https://firebase.com/blog/2014-05-01-using-firebase-with-react.html) on using Firebase with React and check out our [live Todo app demo](https://reactfiretodoapp.firebaseapp.com/) to get started!
 
-Usage
------
+## Usage
+
 The ReactFireMixin can be added to you project in two ways:
 
 * Manually copy ReactFireMixin.js from GitHub to you local directory.
@@ -21,8 +20,8 @@ To use the ReactFireMixin in a React component, add it to the component's mixins
       ...
     });
 
-API Reference
--------------
+## API Reference
+
 ###bindAsArray(firebaseRef, bindVar)
 
 Creates a binding between Firebase and the inputted bind variable as an array. The Firebase
@@ -45,6 +44,24 @@ with that Firebase reference.
 
     this.unbind("items");
 
-License
--------
-[MIT](http://firebase.mit-license.org)
+## Contributing
+
+Interested in manually debugging from source or submitting a pull request? Follow the steps
+below.
+
+### Install Dependencies
+
+```bash
+$ git clone https://github.com/firebase/ReactFire.git       # clone this repository
+$ npm install                                               # install local NPM build / test dependencies
+```
+
+### Compile
+
+```bash
+$ gulp
+```
+
+## License
+
+[Firebase MIT License](http://firebase.mit-license.org)
