+++ bower.json
@@ -1,12 +1,12 @@
 {
   "name": "ReactFire",
-  "version": "0.1.2",
+  "version": "0.1.3",
   "homepage": "https://github.com/firebase/ReactFire",
   "authors": [
     "jacobawenger <jacob@firebase.com>"
   ],
   "description": "Firebase mixin for ReactJS",
-  "main": "ReactFireMixin.js",
+  "main": "js/ReactFireMixin.min.js",
   "keywords": [
     "react",
     "firebase"
@@ -15,6 +15,8 @@
   "ignore": [
     "**/.*",
     "./CHANGELOG.md",
+    "./gulpfile.js",
+    "./package.json",
     "examples",
     "node_modules",
     "bower_components"
