+++ examples/todoApp/bower.json
@@ -23,6 +23,6 @@
   "dependencies": {
     "react": "~0.10.0",
     "firebase": "~1.0.12",
-    "ReactFire": "~0.1.2"
+    "ReactFire": "~0.1.3"
   }
 }
