+++ examples/todoApp/index.html
@@ -11,7 +11,7 @@
     <script src="bower_components/firebase/firebase.js"></script>
 
     <!-- ReactFireMixin -->
-    <script src="bower_components/ReactFire/ReactFireMixin.js"></script>
+    <script src="bower_components/ReactFire/js/ReactFireMixin.min.js"></script>
 
     <!-- Custom JS -->
     <script type="text/jsx" src="js/todoAppOriginal.js"></script>
