+++ package.json
@@ -1,0 +1,30 @@
+{
+  "name": "ReactFire",
+  "description": "Firebase mixin for ReactJS",
+  "main": "ReactFireMixin.js",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/firebase/ReactFire.git"
+  },
+  "bugs": {
+    "url": "https://github.com/firebase/ReactFire/issues"
+  },
+  "private": true,
+  "licenses": [
+    {
+      "type": "MIT",
+      "url": "http://firebase.mit-license.org/"
+    }
+  ],
+  "version": "0.1.3",
+  "dependencies": {
+    "firebase": "~1.0.11"
+  },
+  "devDependencies": {
+    "gulp": "^3.5.6",
+    "gulp-uglify": "^0.2.1",
+    "gulp-jshint": "^1.5.1",
+    "gulp-ext-replace": "0.0.5",
+    "jshint-stylish": "^0.2.0"
+  }
+}
